Fixed issues in getting user information with new structure

+++ server/PlayLingua.Domain/Entities/Invitation.cs
@@ -6,8 +6,8 @@ namespace PlayLingua.Domain.Entities
     {
         public long Id { get; set; }
         public string Email { get; set; }
-        public long TargetLanguageId { get; set; }
-        public long BaseLanguageId { get; set; }
+        public int TargetLanguageId { get; set; }
+        public int BaseLanguageId { get; set; }
         public long? BookId { get; set; }
         public long? ChapterId { get; set; }
         public string GeneratedLink { get; set; }
